Split out ConfigRepoVm and ReposListVM
