+++ server/webapp/WEB-INF/rails/webpack/views/config_repos/config_repos_list.js.msx
@@ -18,7 +18,7 @@ const m = require("mithril");
 const _ = require("lodash");
 const f = require("helpers/form_helper");
 
-const ReposList = {
+const ConfigReposList = {
   view(vnode) {
     const vm = vnode.attrs.vm;
 
@@ -139,4 +139,4 @@ const Field = {
   }
 };
 
-module.exports = ReposList;
+module.exports = ConfigReposList;
