+++ server/webapp/WEB-INF/rails/webpack/views/config_repos/models/repos_list_vm.js
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-const Stream      = require("mithril/stream");
-const _           = require("lodash");
-const Materials   = require("models/config_repos/materials");
-const Validatable = require("models/mixins/validatable_mixin");
-const PluginInfos = require("models/shared/plugin_infos");
+const Stream       = require("mithril/stream");
+const _            = require("lodash");
+const Dfr          = require("jquery").Deferred;
+const PluginInfos  = require("models/shared/plugin_infos");
+const ConfigRepoVM = require("views/config_repos/models/config_repo_vm");
 
 function ReposListVM(model) {
   const repos = Stream([]);
@@ -55,56 +55,6 @@ function ReposListVM(model) {
 
   this.errors = Stream([]);
   this.repos = repos;
-}
-
-function ConfigRepoVM(data) {
-  this.id = Stream();
-  this.pluginId = Stream();
-  this.type = Stream();
-  this.attributes = Stream();
-  this.configuration = Stream();
-  this.etag = Stream(null);
-
-  Validatable.call(this, { errors: {} });
-
-  this.validatePresenceOf("id");
-  this.validatePresenceOf("pluginId");
-  this.validatePresenceOf("type");
-
-  this.allowSave = () => {
-    // intentionally not inlined with `&&` so that we run all validations every time
-    const parentValid = this.isValid();
-    const childValid = this.attributes().isValid();
-    return parentValid && childValid;
-  };
-
-  this.initialize = (data) => {
-    this.id(data.id);
-    this.pluginId(data.plugin_id);
-    this.type(data.material.type);
-    this.attributes(Materials.get(this.type(), data));
-    this.configuration(data.configuration || []);
-  };
-
-  this.initialize(data);
-
-  this.toJSON = () => {
-    return {
-      id: this.id(),
-      plugin_id: this.pluginId(), // eslint-disable-line camelcase
-      material: {
-        type: this.type(),
-        attributes: this.attributes().toJSON()
-      },
-      configuration: _.cloneDeep(this.configuration())
-    };
-  };
-
-  this.clone = () => {
-    const cloned = new ConfigRepoVM(this.toJSON());
-    cloned.etag(this.etag());
-    return cloned;
-  };
 }
 
 // Mixins
